fix: finalize backing history
